+++ frontend/app/index.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { useRouter } from "expo-router";
 import AsyncStorage from "@react-native-async-storage/async-storage";
-import LoginScreen from "./screens/LoginScreen";
+import LoginScreen from "./screens/Auth/LoginScreen";
 import WelcomeScreen from "./screens/WelcomeScreen";
 import { PlayerContextProvider } from './screens/PlayerContext'; // Import PlayerContextProvider
 import * as StorageService from "./services/StorageService";
