+++ frontend/app/screens/Auth/LoginScreen.tsx
@@ -12,8 +12,8 @@ import { useRouter } from "expo-router";
 import { MaterialIcons } from "@expo/vector-icons";
 import { CheckBox } from "react-native-elements";
 import AsyncStorage from "@react-native-async-storage/async-storage";
-import * as StorageService from "../services/StorageService";
-import UserPool from "../services/UserPool";
+import * as StorageService from "../../services/StorageService";
+import UserPool from "../../services/UserPool";
 import { AuthenticationDetails, CognitoUser } from "amazon-cognito-identity-js";
 import { Ionicons } from "@expo/vector-icons";
 import axios from "axios";
@@ -52,7 +52,7 @@ const LoginScreen: React.FC = () => {
         }
 
         // POST request to backend
-        fetch("http://10.32.253.158:3000/auth/login", {
+        fetch("http://localhost:3000/auth/login", {
           method: "POST",
           headers: {
             "Content-Type": "application/json",
