+++ frontend/app/screens/Auth/RegisterStreaming.tsx
