+++ frontend/app/screens/rooms/RoomDetails.tsx
@@ -2,12 +2,12 @@ import React, { useState } from 'react';
 import { View, Text, TextInput, Switch, TouchableOpacity, Dimensions, ScrollView, Image, Button, Alert } from 'react-native';
 import { useLocalSearchParams, useRouter } from 'expo-router';
 import * as ImagePicker from 'expo-image-picker';
-import { RoomDetailsProps } from '../models/roomdetails';
-import { RoomDto } from '../../api-client';
+import { RoomDetailsProps } from '../../models/roomdetails';
+import { RoomDto } from '../../../api-client';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import * as StorageService from "./../services/StorageService"; // Import StorageService
+import * as StorageService from "../../services/StorageService"; // Import StorageService
 import AWS from 'aws-sdk';
-import uploadImage from '../services/ImageUpload';
+import uploadImage from '../../services/ImageUpload';
 
 const AWS_ACCESS_KEY_ID = process.env.AWS_ACCESS_KEY_ID;
 if (!AWS_ACCESS_KEY_ID) {
