+++ frontend/tests/ChatListScreen.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer, { act } from 'react-test-renderer';
-import ChatListScreen from '../app/screens/ChatListScreen';
+import ChatListScreen from '../app/screens/messaging/ChatListScreen';
 import { useRouter } from 'expo-router';
 import { TouchableOpacity, TextInput } from 'react-native';
