+++ frontend/tests/ChatScreen.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer, { act } from 'react-test-renderer';
-import ChatScreen from '../app/screens/ChatScreen';
+import ChatScreen from '../app/screens/messaging/ChatScreen';
 import { useRouter, useLocalSearchParams } from 'expo-router';
 import { TextInput, TouchableOpacity } from 'react-native';
